添加时间和价格范围，预约管理

+++ server/src/main/java/com/leoman/areaRange/service/impl/AreaRangeServiceImpl.java
@@ -37,7 +37,7 @@ public class AreaRangeServiceImpl extends GenericManagerImpl<AreaRange,AreaRange
         }else if(areaRange.getAreaTo() == null || StringUtils.isEmpty(areaRange.getAreaTo())){
             str = areaRange.getAreaFrom() + "㎡以上";
         }else{
-            str = areaRange.getAreaFrom().floatValue() +"-"+ areaRange.getAreaTo().floatValue() + "㎡";
+            str = areaRange.getAreaFrom() +"-"+ areaRange.getAreaTo() + "㎡";
         }
         areaRange.setName(str);
